refactor: inputFieldStyles.ts 디렉터리 변경
