+++ src/feature/CardRegister/components/inputs/CvcField/CvcField.tsx
@@ -2,7 +2,7 @@ import Input from '@/common/components/Input/Input';
 import styled from 'styled-components';
 import type {CardRegisterInputProps} from '../shared.types';
 import InputContainer from '../InputContainer/InputContainer';
-import {FieldErrorMessage, FieldLayout} from '../../styles/inputFieldStyles';
+import {FieldErrorMessage, FieldLayout} from '../styles/inputFieldStyles';
 
 type CvcFieldProps = {
   inputProps: CardRegisterInputProps;
